hello: Clean up the hello world example changes
Add a lot of comments, replace hand-coded xml functions with standard
one, use clixon_debug() instead of printf in a lot of places, and
comment out the printing of the XML trees by default.

Also show an example of both parsing the source and target trees or
parsing the add,delete,change vectors.

Signed-off-by: Corey Minyard <minyard@acm.org>